backend/requirements.txt
asgiref==3.2.5
avro-python3==1.9.1
cachetools==3.1.1
certifi==2019.11.28
crcmod==1.7
Django==3.0.4
django-filter==2.2.0
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
docopt==0.6.2
fastavro==0.21.24
fasteners==0.15
future==0.18.2
google-apitools==0.5.28
google-cloud-bigtable==1.0.0
google-cloud-datastore==1.7.4
google-cloud-pubsub==1.0.2
googleapis-common-protos==1.6.0
grpc-google-iam-v1==0.12.3
grpcio==1.25.0
hdfs==2.5.8
httplib2==0.12.0
Markdown==3.2.1
mock==2.0.0
monotonic==1.5
mysqlclient==1.4.6
numpy==1.18.1
oauth2client==3.0.0
pandas>=1.1.0
pbr==5.4.3
protobuf==3.10.0
pyasn1==0.4.7
pyasn1-modules==0.2.7
pydot==1.4.1
PyJWT==1.7.1
pymongo==3.9.0
pyparsing==2.4.4
python-dateutil==2.8.1
pytz==2019.3
rsa==4.0
six==1.14.0
sqlparse==0.3.1
text-unidecode==1.3
wincertstore==0.2
scikit-learn
sqlalchemy
